requirements.txt
streamlit
pandas
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=0.8.11
httpx>=0.27.0
